Fix OnUpdated action bug from Timer
If the next timer from GetNextTimer is null, the for loop continues and skips further execution. Insert the OnUpdated invocation before continuing the loop.

If the next timer is done with the same end time as the current timer, the current timer does not invoke its OnComplete action because two statements (from the else condition) are placed on the same line as the if statement. The loop is continued and skips further execution because it's out of scope.

- Update package version